Add responsive toast types
